Reformat database with new schema
db_model was changed to reflect the new "single skill table" format.
jobfind.db also reflects these changes. If you had data in the old tables,
it is now gone and must be re-entered (i.e., if you had an account it must
be registered again).

Additionally, the User table now has a column for "isAdmin" and "isFaculty"
which are booleans (1 or 0 in sqlite3). This eliminates the need for a
separate admin and faculty tables.