Create skills list table in database
The database now has a table LIST which contains a list of all of the
skills listed EXACTLY as they are defined in the SKILLS table. If you
remake the database, be sure to run this from the command line
cat skills.sql | sqlite3 jobfind.db
this will re-add the skills to the database.